# gateway/docs/swagger.yaml
basePath: /
...
host: localhost:8080
info:
  contact:
    email: support@erp-system.com
    name: ERP System Support
  description: Enterprise Resource Planning System API Gateway. Provides unified REST
    API access to Inventory, Order, Finance, and HR microservices.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: ERP System API
  version: "1.0"
paths:
...
  /api/v1/inventory/categories:
    get:
      description: Retrieve all product categories
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: List all categories
      tags:
      - Inventory - Categories
    post:
      consumes:
      - application/json
      description: Create a new product category
      parameters:
      - description: Category data (name, description)
        in: body
        name: category
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create a new category
      tags:
      - Inventory - Categories
  /api/v1/inventory/products:
    get:
      description: Retrieve a paginated list of products with optional filters
      parameters:
      - default: "1"
        description: Page number
        in: query
        name: page
        type: string
      - default: "20"
        description: Items per page
        in: query
        name: page_size
        type: string
      - description: Filter by category
        in: query
        name: category_id
        type: string
      - default: ACTIVE
        description: Filter by status
        in: query
        name: status
        type: string
      - description: Search query
        in: query
        name: q
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: List all products
      tags:
      - Inventory - Products
    post:
      consumes:
      - application/json
      description: Create a new product in the inventory system
      parameters:
      - description: Product data (name, sku, category_id, price, etc.)
        in: body
        name: product
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: Product created successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request body
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create a new product
      tags:
      - Inventory - Products
  /api/v1/inventory/products/{id}:
    get:
      description: Retrieve a single product by its ID
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get a product by ID
      tags:
      - Inventory - Products
    put:
      consumes:
      - application/json
      description: Update an existing product by its ID
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: string
      - description: Updated product data
        in: body
        name: product
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Update a product
      tags:
      - Inventory - Products
  /api/v1/inventory/stock/{productId}:
    get:
      description: Retrieve stock levels across all warehouses for a given product
      parameters:
      - description: Product ID
        in: path
        name: productId
        required: true
        type: string
      - description: Filter by warehouse
        in: query
        name: warehouse_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get stock levels for a product
      tags:
      - Inventory - Stock
  /api/v1/inventory/stock/adjust:
    post:
      consumes:
      - application/json
      description: Manually adjust stock quantity for a product in a warehouse
      parameters:
      - description: Stock adjustment data (product_id, warehouse_id, quantity, reason)
        in: body
        name: adjustment
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Adjust stock quantity
      tags:
      - Inventory - Stock
  /api/v1/inventory/stock/alerts:
    get:
      description: Retrieve products that are below their minimum stock threshold
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get low stock alerts
      tags:
      - Inventory - Stock
  /api/v1/inventory/stock/movements:
    get:
      description: Retrieve a history of all stock movements (adjustments, transfers,
        etc.)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: List stock movements
      tags:
      - Inventory - Stock
  /api/v1/inventory/stock/reserve:
    post:
      consumes:
      - application/json
      description: Reserve stock quantity for a pending order
      parameters:
      - description: Reservation data (product_id, warehouse_id, quantity, order_id)
        in: body
        name: reservation
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Reserve stock for an order
      tags:
      - Inventory - Stock
  /api/v1/inventory/stock/transfer:
    post:
      consumes:
      - application/json
      description: Transfer stock from one warehouse to another
      parameters:
      - description: Transfer data (product_id, from_warehouse, to_warehouse, quantity)
        in: body
        name: transfer
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Transfer stock between warehouses
      tags:
      - Inventory - Stock
  /api/v1/inventory/warehouses:
    post:
      consumes:
      - application/json
      description: Create a new warehouse location
      parameters:
      - description: Warehouse data (name, code, address)
        in: body
        name: warehouse
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create a new warehouse
      tags:
      - Inventory - Warehouses
...
securityDefinitions:
  BearerAuth:
    description: Enter your JWT token with the Bearer prefix, e.g. "Bearer eyJhbGciOiJI..."
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...